@@ -1,7 +1,8 @@
 import { Injectable, NgZone, ApplicationRef, InjectionToken, Inject, Optional } from '@angular/core';
-import { Observable } from 'rxjs';
-import { first, tap } from 'rxjs/operators';
+import { Observable, Subscription, from } from 'rxjs';
+import { first, tap, map, shareReplay, switchMap } from 'rxjs/operators';
 import { performance } from 'firebase/app';
+import { FirebaseApp } from '@angular/fire';
 
 export const AUTOMATICALLY_TRACE_CORE_NG_METRICS = new InjectionToken<boolean>('angularfire2.performance.auto_trace');
 export const INSTRUMENTATION_ENABLED = new InjectionToken<boolean>('angularfire2.performance.instrumentationEnabled');
@@ -18,9 +19,10 @@ export type TraceOptions = {
 @Injectable()
 export class AngularFirePerformance {
   
-  performance: performance.Performance;
+  performance: Observable<performance.Performance>;
 
   constructor(
+    app: FirebaseApp,
     @Optional() @Inject(AUTOMATICALLY_TRACE_CORE_NG_METRICS) automaticallyTraceCoreNgMetrics:boolean|null,
     @Optional() @Inject(INSTRUMENTATION_ENABLED) instrumentationEnabled:boolean|null,
     @Optional() @Inject(DATA_COLLECTION_ENABLED) dataCollectionEnabled:boolean|null,
@@ -28,66 +30,94 @@ export class AngularFirePerformance {
     private zone: NgZone
   ) {
     
-    this.performance = zone.runOutsideAngular(() => performance());
+    // @ts-ignore zapping in the UMD in the build script
+    const requirePerformance = from(import('firebase/performance'));
 
-    if (instrumentationEnabled == false) { this.performance.instrumentationEnabled = false }
-    if (dataCollectionEnabled == false) { this.performance.dataCollectionEnabled = false }
-    
+    this.performance = requirePerformance.pipe(
+      // SEMVER while < 6 need to type, drop next major
+      map(() => zone.runOutsideAngular(() => <performance.Performance>app.performance())),
+      tap(performance => {
+        if (instrumentationEnabled == false) { performance.instrumentationEnabled = false }
+        if (dataCollectionEnabled == false) { performance.dataCollectionEnabled = false }
+      }),
+      shareReplay(1)
+    );
+
     if (automaticallyTraceCoreNgMetrics != false) {
 
       // TODO determine more built in metrics
       appRef.isStable.pipe(
-        this.traceUntilLast('isStable'),
-        first(it => it)
+        first(it => it),
+        this.traceUntilComplete('isStable')
       ).subscribe();
 
     }
 
   }
 
-  trace$ = (name:string, options?: TraceOptions) => new Observable<void>(emitter => 
-    this.zone.runOutsideAngular(() => {
-      const trace = this.performance.trace(name);
-      options && options.metrics && Object.keys(options.metrics).forEach(metric => {
-        trace.putMetric(metric, options!.metrics![metric])
-      });
-      options && options.attributes && Object.keys(options.attributes).forEach(attribute => {
-        trace.putAttribute(attribute, options!.attributes![attribute])
-      });
-      const attributeSubscriptions = options && options.attribute$ ? Object.keys(options.attribute$).map(attribute =>
-        options!.attribute$![attribute].subscribe(next => trace.putAttribute(attribute, next))
-      ) : [];
-      const metricSubscriptions = options && options.metric$ ? Object.keys(options.metric$).map(metric =>
-        options!.metric$![metric].subscribe(next => trace.putMetric(metric, next))
-      ) : [];
-      const incrementOnSubscriptions = options && options.incrementMetric$ ? Object.keys(options.incrementMetric$).map(metric =>
-        options!.incrementMetric$![metric].subscribe(next => trace.incrementMetric(metric, next || undefined))
-      ) : [];
-      emitter.next(trace.start());
-      return { unsubscribe: () => {
-        trace.stop();
-        metricSubscriptions.forEach(m => m.unsubscribe());
-        incrementOnSubscriptions.forEach(m => m.unsubscribe());
-        attributeSubscriptions.forEach(m => m.unsubscribe());
-      }};
-    })
-  );
+  trace$ = (name:string, options?: TraceOptions) =>
+    this.performance.pipe(
+      switchMap(performance =>
+        new Observable<void>(emitter =>
+          this.zone.runOutsideAngular(() => {
+            const trace = performance.trace(name);
+            options && options.metrics && Object.keys(options.metrics).forEach(metric => {
+              trace.putMetric(metric, options!.metrics![metric])
+            });
+            options && options.attributes && Object.keys(options.attributes).forEach(attribute => {
+              trace.putAttribute(attribute, options!.attributes![attribute])
+            });
+            const attributeSubscriptions = options && options.attribute$ ? Object.keys(options.attribute$).map(attribute =>
+              options!.attribute$![attribute].subscribe(next => trace.putAttribute(attribute, next))
+            ) : [];
+            const metricSubscriptions = options && options.metric$ ? Object.keys(options.metric$).map(metric =>
+              options!.metric$![metric].subscribe(next => trace.putMetric(metric, next))
+            ) : [];
+            const incrementOnSubscriptions = options && options.incrementMetric$ ? Object.keys(options.incrementMetric$).map(metric =>
+              options!.incrementMetric$![metric].subscribe(next => trace.incrementMetric(metric, next || undefined))
+            ) : [];
+            emitter.next(trace.start());
+            return { unsubscribe: () => {
+              trace.stop();
+              metricSubscriptions.forEach(m => m.unsubscribe());
+              incrementOnSubscriptions.forEach(m => m.unsubscribe());
+              attributeSubscriptions.forEach(m => m.unsubscribe());
+            }};
+          })
+        )
+      )
+    );
 
-  traceUntil = <T=any>(name:string, test: (a:T) => boolean, options?: TraceOptions) => (source$: Observable<T>) => {
+  traceUntil = <T=any>(name:string, test: (a:T) => boolean, options?: TraceOptions & { orComplete?: boolean }) => (source$: Observable<T>) => new Observable<T>(subscriber => {
     const traceSubscription = this.trace$(name, options).subscribe();
     return source$.pipe(
-      tap(a => { if (test(a)) { traceSubscription.unsubscribe() }})
-    )
-  };
+      tap(
+        a  => test(a) && traceSubscription.unsubscribe(),
+        () => {},
+        () => options && options.orComplete && traceSubscription.unsubscribe()
+      )
+    ).subscribe(subscriber);
+  });
 
-  traceWhile = <T=any>(name:string, test: (a:T) => boolean, options?: TraceOptions) => (source$: Observable<T>) => {
-    const traceSubscription = this.trace$(name, options).subscribe();
+  traceWhile = <T=any>(name:string, test: (a:T) => boolean, options?: TraceOptions & { orComplete?: boolean}) => (source$: Observable<T>) => new Observable<T>(subscriber => {
+    let traceSubscription: Subscription|undefined;
     return source$.pipe(
-      tap(a => { if (!test(a)) { traceSubscription.unsubscribe() }})
-    )
-  };
+      tap(
+        a  => {
+          if (test(a)) {
+            traceSubscription = traceSubscription || this.trace$(name, options).subscribe();
+          } else {
+            traceSubscription && traceSubscription.unsubscribe();
+            traceSubscription = undefined;
+          }
+        },
+        () => {},
+        () => options && options.orComplete && traceSubscription && traceSubscription.unsubscribe()
+      )
+    ).subscribe(subscriber);
+  });
 
-  traceUntilLast= <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => {
+  traceUntilComplete = <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => new Observable<T>(subscriber => {
     const traceSubscription = this.trace$(name, options).subscribe();
     return source$.pipe(
       tap(
@@ -95,10 +125,10 @@ export class AngularFirePerformance {
         () => {},
         () => traceSubscription.unsubscribe()
       )
-    )
-  };
+    ).subscribe(subscriber);
+  });
 
-  traceUntilFirst = <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => {
+  traceUntilFirst = <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => new Observable<T>(subscriber => {
     const traceSubscription = this.trace$(name, options).subscribe();
     return source$.pipe(
       tap(
@@ -106,10 +136,10 @@ export class AngularFirePerformance {
         () => {},
         () => {}
       )
-    )
-  };
+    ).subscribe(subscriber);
+  });
 
-  trace = <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => {
+  trace = <T=any>(name:string, options?: TraceOptions) => (source$: Observable<T>) => new Observable<T>(subscriber => {
     const traceSubscription = this.trace$(name, options).subscribe();
     return source$.pipe(
       tap(
@@ -117,7 +147,7 @@ export class AngularFirePerformance {
         () => {},
         () => traceSubscription.unsubscribe()
       )
-    )
-  };
+    ).subscribe(subscriber);
+  });
 
 }
