@@ -3,6 +3,9 @@ import { FirebaseApp } from '@angular/fire';
 import { AngularFireDatabase } from '@angular/fire/database';
 import { AngularFireAuth } from '@angular/fire/auth';
 import { AngularFirestore } from '@angular/fire/firestore';
+import { AngularFireStorage } from '@angular/fire/storage';
+import { AngularFireMessaging } from '@angular/fire/messaging';
+import { AngularFireFunctions } from '@angular/fire/functions';
 
 @Component({
   selector: 'app-root',
@@ -32,8 +35,11 @@ export class AppComponent {
     private readonly app: FirebaseApp,
     private readonly db: AngularFireDatabase,
     private readonly auth: AngularFireAuth,
-    private readonly afStore: AngularFirestore
+    private readonly afStore: AngularFirestore,
+    private readonly storage: AngularFireStorage,
+    private readonly messaging: AngularFireMessaging,
+    private readonly functions: AngularFireFunctions
   ) {
-    console.log(app, db, auth, afStore);
+    console.log(app, db, auth, afStore, storage, messaging, functions);
   }
 }
